Refactor download util.
