+++ app/src/main/java/tech/ula/utils/Asset.kt
@@ -1,0 +1,6 @@
+package tech.ula.utils
+
+data class Asset(
+    val name: String,
+    val location: String
+)
