+++ app/src/main/java/tech/ula/utils/FlowUtility.kt
@@ -1,0 +1,10 @@
+package tech.ula.utils
+
+class FlowUtility {
+    fun largeAssetRequiredAndNoWifi(): Boolean {
+        val filesystemIsPresent = session.isExtracted || filesystem.isDownloaded
+        return !(filesystemIsPresent || wifiIsEnabled())
+    }
+
+    fun
+}
