+++ app/src/main/java/tech/ula/utils/ProgressBarUpdater.kt
@@ -1,0 +1,6 @@
+package tech.ula.utils
+
+data class ProgressBarUpdates(
+    var step: String,
+    var details: String
+)
