+++ app/src/main/java/tech/ula/utils/SessionInstaller.kt
@@ -1,0 +1,3 @@
+package tech.ula.utils
+
+class SessionInstaller
